add interaction with postgres and chart proper temperature

+++ bff/types/datapoint.go
@@ -2,7 +2,7 @@ package types
 
 type TemperaturePoint struct {
 	Timestamp    float64
-	TemperatureC int
+	TemperatureC float64
 	Name         string
 }
